+++ compiler/frontend/beam/src/main/java/org/apache/nemo/compiler/frontend/beam/transform/DoFnTransform.java
@@ -21,23 +21,20 @@ package org.apache.nemo.compiler.frontend.beam.transform;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.apache.beam.sdk.values.PCollectionView;
 import org.apache.beam.sdk.values.TupleTag;
 import org.apache.beam.sdk.values.WindowingStrategy;
 import org.apache.nemo.common.ir.OutputCollector;
 import org.apache.nemo.common.punctuation.Watermark;
-import org.apache.nemo.compiler.frontend.beam.SideInputElement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
 /**
- * DoFn transform implementation.
+ * DoFn transform implementation when there is no side input.
  *
  * @param <InputT> input type.
  * @param <OutputT> output type.
@@ -45,18 +42,8 @@ import java.util.Map;
 public final class DoFnTransform<InputT, OutputT> extends AbstractDoFnTransform<InputT, InputT, OutputT> {
   private static final Logger LOG = LoggerFactory.getLogger(DoFnTransform.class.getName());
 
-  private List<WindowedValue<InputT>> curPushedBacks;
-  private long curPushedBackWatermark; // Long.MAX_VALUE when no pushed-back exists.
-  private long curInputWatermark;
-  private long curOutputWatermark;
-
-  private final boolean noSideInput;
-
   /**
    * DoFnTransform Constructor.
-   *
-   * @param doFn    doFn.
-   * @param options Pipeline options.
    */
   public DoFnTransform(final DoFn<InputT, OutputT> doFn,
                        final Coder<InputT> inputCoder,
@@ -68,11 +55,9 @@ public final class DoFnTransform<InputT, OutputT> extends AbstractDoFnTransform<
                        final PipelineOptions options) {
     super(doFn, inputCoder, outputCoders, mainOutputTag,
       additionalOutputTags, windowingStrategy, sideInputs, options);
-    this.curPushedBacks = new ArrayList<>();
-    this.curPushedBackWatermark = Long.MAX_VALUE;
-    this.curInputWatermark = Long.MIN_VALUE;
-    this.curOutputWatermark = Long.MIN_VALUE;
-    this.noSideInput = sideInputs.isEmpty();
+    if (!sideInputs.isEmpty()) {
+      throw new IllegalStateException(sideInputs.toString());
+    }
   }
 
   @Override
@@ -83,85 +68,19 @@ public final class DoFnTransform<InputT, OutputT> extends AbstractDoFnTransform<
   @Override
   public void onData(final Object data) {
     // Do not need any push-back logic.
-    if (noSideInput) {
-      checkAndInvokeBundle();
-      final WindowedValue<InputT> mainInputElement = (WindowedValue<InputT>) data;
-      getDoFnRunner().processElement(mainInputElement);
-      checkAndFinishBundle(false);
-      return;
-    }
-
-    // Need to distinguish side/main inputs and push-back main inputs.
-    if (data instanceof SideInputElement) {
-      // This element is a Side Input
-      final SideInputElement sideInputElement = (SideInputElement) data;
-      // TODO #287: Consider Explicit Multi-Input IR Transform
-
-      // Flush out any current bundle-related states in the DoFn,
-      // as this sideinput may trigger the processing of pushed-back data.
-      checkAndFinishBundle(true); // forced
-
-      checkAndInvokeBundle();
-      final PCollectionView view = getSideInputs().get(sideInputElement.getSideInputIndex());
-      final WindowedValue sideInputData = sideInputElement.getSideInputValue();
-      getSideInputHandler().addSideInputValue(view, sideInputData);
-
-      // With the new side input added, we may be able to process some pushed-back elements.
-      final List<WindowedValue<InputT>> pushedBackAgain = new ArrayList<>();
-      long pushedBackAgainWatermark = Long.MAX_VALUE;
-      for (WindowedValue<InputT> curPushedBack : curPushedBacks) {
-        final Iterable<WindowedValue<InputT>> pushedBack =
-          getPushBackRunner().processElementInReadyWindows(curPushedBack);
-        for (final WindowedValue<InputT> wv : pushedBack) {
-          pushedBackAgainWatermark = Math.min(pushedBackAgainWatermark, wv.getTimestamp().getMillis());
-          pushedBackAgain.add(wv);
-        }
-      }
-      curPushedBacks = pushedBackAgain;
-      curPushedBackWatermark = pushedBackAgainWatermark;
-      checkAndFinishBundle(false);
-
-      // See if we can emit a new watermark, as we may have processed some pushed-back elements
-      onWatermark(new Watermark(curInputWatermark));
-    } else {
-      // This element is the Main Input
-      final WindowedValue<InputT> mainInputElement = (WindowedValue<InputT>) data;
-      checkAndInvokeBundle();
-      final Iterable<WindowedValue<InputT>> pushedBack =
-        getPushBackRunner().processElementInReadyWindows(mainInputElement);
-      for (final WindowedValue wv : pushedBack) {
-        curPushedBackWatermark = Math.min(curPushedBackWatermark, wv.getTimestamp().getMillis());
-        curPushedBacks.add(wv);
-      }
-      checkAndFinishBundle(false);
-    }
+    checkAndInvokeBundle();
+    final WindowedValue<InputT> mainInputElement = (WindowedValue<InputT>) data;
+    getDoFnRunner().processElement(mainInputElement);
+    checkAndFinishBundle(false);
   }
 
   @Override
   public void onWatermark(final Watermark watermark) {
-    curInputWatermark = watermark.getTimestamp();
-
-    final long minOfInputAndPushback = Math.min(curInputWatermark, curPushedBackWatermark);
-    if (minOfInputAndPushback > curOutputWatermark) {
-      // Watermark advances!
-      getOutputCollector().emitWatermark(new Watermark(minOfInputAndPushback));
-      curOutputWatermark = minOfInputAndPushback;
-    }
-
-    if (watermark.getTimestamp() >= BoundedWindow.TIMESTAMP_MAX_VALUE.getMillis()) {
-      hardFlushAllPushedbacks();
-    }
-  }
-
-  private void hardFlushAllPushedbacks() {
-    // Instead of using the PushBackRunner, we directly use the DoFnRunner to not wait for sideinputs.
-    curPushedBacks.forEach(wv -> getDoFnRunner().processElement(wv));
-    curPushedBacks.clear();
+    getOutputCollector().emitWatermark(watermark);
   }
 
   @Override
   protected void beforeClose() {
-    hardFlushAllPushedbacks();
   }
 
   @Override
